Remove email verification modal
