+++ frontend/src/components/pages/Home.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { NavLink, withRouter } from "react-router-dom";
-import MenteeVerificationModal from "../MenteeVerificationModal";
+import LoginVerificationModal from "../MenteeVerificationModal";
 import "../css/Home.scss";
 import Logo from "../../resources/logo.png";
 import Health from "../../resources/focus-for-health.svg";
@@ -13,7 +13,7 @@ function Home({ history }) {
           <h1 className="home-header">Welcome to MENTEE</h1>
           <p className="home-text">Find a global mentor now...</p>
           <br />
-          <MenteeVerificationModal
+          <LoginVerificationModal
             content={<b>Find a Mentor</b>}
             theme="dark"
             onVerified={() => {
